Extract tooltip implementation into base OsuSliderBar.cs from NormalSliderBar.cs
